First iteration of the compile command
It needs tests, some prettying up, and to look into the @todo
items that I left.

However, it is working in most cases. The only case that does not
seem to work right now is that I seem to be using the same logic
as #require when searching for a gem, so gems that do not put
requirable files in the top level of their require paths will
not work (e.g. minitest has to be required using minitest/unit).